+++ src/main/java/roomescape/exception/BadRequestException.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.exception;
 
 public class BadRequestException extends RuntimeException {
