+++ src/main/java/roomescape/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.exception;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
