+++ src/main/java/roomescape/exception/ReservationNotFoundException.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.exception;
 
 public class ReservationNotFoundException extends RuntimeException {
