+++ src/main/java/roomescape/repository/InMemoryReservationRepository.java
@@ -1,6 +1,7 @@
-package roomescape;
+package roomescape.repository;
 
 import org.springframework.stereotype.Repository;
+import roomescape.domain.Reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
