+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,6 +1,11 @@
-package roomescape;
+package roomescape.service;
 
 import org.springframework.stereotype.Service;
+import roomescape.domain.Reservation;
+import roomescape.dto.ReservationRequest;
+import roomescape.exception.BadRequestException;
+import roomescape.exception.ReservationNotFoundException;
+import roomescape.repository.ReservationRepository;
 
 import java.time.LocalDateTime;
 import java.util.List;
