+++ src/test/java/roomescape/ReservationServiceTest.java
@@ -2,6 +2,13 @@ package roomescape;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import roomescape.domain.Reservation;
+import roomescape.dto.ReservationRequest;
+import roomescape.exception.BadRequestException;
+import roomescape.exception.ReservationNotFoundException;
+import roomescape.repository.InMemoryReservationRepository;
+import roomescape.repository.ReservationRepository;
+import roomescape.service.ReservationService;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
